# api/docs/swagger.yaml
definitions:
...
  schema.AstraEvent:
    properties:
      activity_name:
        type: string
      capacity:
        type: number
      current_state:
        type: string
      end_date:
        type: string
      meeting_type:
        type: string
      not_allowed_usage_mask:
        type: number
      start_date:
        type: string
      usage_color:
        type: string
    type: object
...
  schema.MultiBuildingEvents-schema_AstraEvent:
    properties:
      buildings:
        items:
          $ref: '#/definitions/schema.SingleBuildingEvents-schema_AstraEvent'
        type: array
      date:
        type: string
    type: object
...
  schema.MultiBuildingEvents-schema_SectionWithTime:
    properties:
      buildings:
        items:
          $ref: '#/definitions/schema.SingleBuildingEvents-schema_SectionWithTime'
        type: array
      date:
        type: string
    type: object
...
  schema.RoomEvents-schema_AstraEvent:
    properties:
      events:
        items:
          $ref: '#/definitions/schema.AstraEvent'
        type: array
      room:
        type: string
    type: object
...
  schema.RoomEvents-schema_SectionWithTime:
    properties:
      events:
        items:
          $ref: '#/definitions/schema.SectionWithTime'
        type: array
      room:
        type: string
    type: object
...
  schema.SectionWithTime:
    properties:
      end_time:
        type: string
      section:
        type: string
      start_time:
        type: string
    type: object
...
  schema.SingleBuildingEvents-schema_AstraEvent:
    properties:
      building:
        type: string
      rooms:
        items:
          $ref: '#/definitions/schema.RoomEvents-schema_AstraEvent'
        type: array
    type: object
...
  schema.SingleBuildingEvents-schema_SectionWithTime:
    properties:
      building:
        type: string
      rooms:
        items:
          $ref: '#/definitions/schema.RoomEvents-schema_SectionWithTime'
        type: array
    type: object
host: api.utdnebula.com
info:
  contact: {}
  description: The developer Nebula Labs API for access to pertinent UT Dallas data
  title: dev-nebula-api
  version: 1.0.0
paths:
  /astra/{date}:
    get:
      description: '"Returns AstraEvent based on the input date"'
      operationId: AstraEvents
      parameters:
      - description: date (ISO format) to retrieve astra events
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All AstraEvents with events on the inputted date
          schema:
            items:
              $ref: '#/definitions/schema.MultiBuildingEvents-schema_AstraEvent'
            type: array
...
  /events/{date}:
    get:
      description: '"Returns all sections with meetings on the specified date"'
      operationId: events
      parameters:
      - description: ISO date of the set of events to get
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All sections with meetings on the specified date
          schema:
            items:
              $ref: '#/definitions/schema.MultiBuildingEvents-schema_SectionWithTime'
            type: array
  /events/{date}/{building}:
    get:
      description: '"Returns all sections with meetings on the specified date in the
        specified building"'
      operationId: eventsByBuilding
      parameters:
      - description: ISO date of the set of events to get
        in: path
        name: date
        required: true
        type: string
      - description: building abbreviation of event locations
        in: path
        name: building
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: All sections with meetings on the specified date in the specified
            building
          schema:
            items:
              $ref: '#/definitions/schema.SingleBuildingEvents-schema_SectionWithTime'
            type: array
...
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: x-api-key
    type: apiKey
  storage_key:
    in: header
    name: x-storage-key
    type: apiKey
swagger: "2.0"
...